requirements.txt
altair==5.5.0
asgiref==3.8.1
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
colorama==0.4.6
Django==5.1.5
djangorestframework==3.15.2
filelock==3.17.0
Flask==3.1.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.25.1
netifaces==0.11.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pefile==2024.8.26
pillow==11.1.0
plotly==6.0.0
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-nmap==0.7.1
pytz==2025.1
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scapy==2.6.1
scikit-learn==1.6.1
scipy==1.15.1
shodan==1.31.0
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
streamlit==1.42.0
tenacity==9.0.0
threadpoolctl==3.5.0
tldextract==5.1.3
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
XlsxWriter==3.2.2
